router: fix panic when downloading devices from the gateway (#334)
The cause of the panic was proto.Merge not behaving as advertised when merging proto.Messages that have equivalent but != protoreflect.Descriptor instances, like we have on the gateway as we use protodesc and dynamicpb packages to create dynamic descriptors and messages.

The fix is to fall back to the slow path in these cases and round trip the merge via proto.Mershal/Unmarshal. proto.Merge is advertised to be equivalent to this anyway.

Fixes SC-1002