history: don't call Slice.Len when it's unlikely to have changed
Calling history.Slice.Len for each page has large performance implications on the application, especially in cases where many pages are being read. In situations where Len is backed by a database call it's generally implemented via a `select count(*) where d>%T AND d<=%T` which can be very slow.

The problem is easy to avoid as the total size for a pagination operation is unlikely to change between page calls, so we solve this by calling Len for the first page and keeping that value in the page token for reuse by the next page calls.

During testing this changed history download from ~60KB/s to ~2MB/s, or to put it another way downloading ~1m records now takes seconds instead of hours.