history: add order_by to history requests
This change updates both the trait specific and history admin apis to allow response order to be specified by the client. The orders we support are limited to time ASC and time DESC, but that's really all we need.

The use case this change enables is fetching the last known value before a given point in time, i.e. what was the last meter reading before 00:00 on the 2nd Feb. This is something that was impossible before and resulted in less than idea UX on UIs, for example charts.

I noticed that the timestamp field of the admin api and the trait apis are called different things: create_time and record_time. I haven't changed these as part of the PR as there were already a lot of things touched by this change. Similarly, I've kept the default Read operation instead of renaming it to ReadAsc.

The order_by field is modelled after the AIP ordering recommendation (https://google.aip.dev/132#ordering) even though we support a single field and a `bool desc` could have worked.