history: fix chained set widening in Slice.Slice calls
Previously, all stores apart from the memstore would incorrectly expand the slice when chaining calls to Slice. i.e. [1,2,3,4].Slice(2, 4).Slice(1, 5) should return the slice [2,3], but would return [1,2,3,4] due to the incorrect widening.

The fix was to make the slice stores use set intersection when Slice is called.

The memstore didn't have this issue because the slicing was performed in the Slice call, the other stores performed the slicing in the Read.