ui/ops/energy: add EnergyHistoryCard widget
The EnergyHistoryCard is a widget that shows stacked energy usage for a number of sources, intended to break down (for example) a building by floors to give a broader overview of energy use where ancestry is known.

This change also updates the ConsumptionCard widget as additional time utils have been written that this component can make use of.