devices: allow the caller to specify a filename for download
In some situations (browser, cross origin, etc) the <a download="filename"> attribute isn't enough to have the file downloaded with the correct filename. In these cases the servers Content-Disposition header takes president which meant it was always download.csv.

I've put an arbitrary filename size limit of 255 for now, just to make sure we aren't acting badly as the filename will get embedded into the response headers when the url is GET.