ui/ops: fix race in useCollection
I've reimplemented the list part of useCollection in terms of the race free useAction. This fixes all the cases where typing too quickly in search boxes would cause old list results to appear on the ui.

It might just be me, but the responsiveness also seems to be way better too.

Fixes SC-985