all: replace grpc.Dial/Context calls with grpc.NewClient
Most of the work here is about also removing the grpc.WithBlock and related options which aren't usable with NewClient.

There were some logic changes needed in the hub packages, but I don't think they'll have any noticeable effects on our deployments.

The surprising thing was in node/remote.go which required a Connect call to build the resolver before the address could be updated. Without this the tests would panic.