auto/bms: allow occupancy to be changed on a schedule (#398)
The schedule works either independently or with the occupancy sensors. Without sensors the occupancy mode will be set to on when within the schedule, and off otherwise. When sensors exist, occupied mode is only ever set to on within the schedule using the normal occupancy sensor logic.

I've added tests for all the occupancy cases, both existing and schedule based, just to make sure I didn't introduce any regressions here.