system/history: fix support for order_by clause (#401)
This looks to have been missed when the feature was added and never tested in the context of a cohort.

I've fixed it by publishing the historypb page functionality in a way it can be used by both trait history APIs and the history admin api. The history system now uses this shared pagination logic along with the proto page token types.