auto/history: refactor to allow pollOnSchedule lifecycle defined by a cron schedule in config
The default behaviour will be as before: without a cron schedule for the history automation, the pull.Fetcher will determine whether polling or pull operations are used to fetch changes from a device trait. However, if a cron schedule is provided, the history automation will use a poll cycle to fetch values on a given schedule, and will ignore duplicate consecutive changes to device traits. This is effort taken to reduce the storage required for a history automation.