auto/history: refactor to allow pollOnSchedule lifecycle defined by a cron schedule in config (#407)
The default behaviour will be as before: without a cron schedule for the history automation, the pull.Fetcher will determine whether polling or pull operations are used to fetch changes from a device trait. However, if a cron schedule is provided, the history automation will use a poll lifecycle to fetch values on a given schedule, and will ignore duplicate consecutive changes to device traits. This is effort taken to reduce the storage required for a history automation.

* util/jsontypes: add a more relaxed cron parser

* auto/history: refactor to allow pollOnSchedule lifecycle defined by a cron schedule in config

* auto/history: default comparator in deduper constructor