account: add roles that map to the legacy hardcoded roles
The legacy named roles 'admin', 'super-admin', 'commissioner', 'operator' and 'viewer' are still used by policies etc. to decide access.
Until we modernise the rest of sc-bos to use the Role Assignments / Permissions model encoded in the accounts database, we need to store these legacy roles as well.

Pre-populate the database with roles to represent these legacy roles.
They are protected so they can't be altered or deleted.

The legacy role system doesn't have the concept of scoped role assignments,
so we forbid the legacy roles to be used in that case.