authn: support user login with accounts database
Login as a user account stored in the accounts database.
Unlike file-based accounts, accounts in the database can be modified from the Ops UI and accounts API.

This login uses the legacy roles system, so the user logging in must be assigned at least one role mapped to a legacy role.