auto/history: refactor to allow pollOnSchedule lifecycle defined by a cron schedule in config and fix history/memory store for tests (#409)
This is a new feature where a schedule can be provided to a history automation.

The default behaviour will be as before: without a cron schedule for the history automation, the pull.Fetcher will determine whether polling or pull operations are used to fetch changes from a device trait. However, if a cron schedule is provided, the history automation will use a poll lifecycle to fetch values on a given schedule, and will ignore duplicate consecutive changes to device traits. This is effort taken to reduce the storage required for a history automation.

Also, fixes a condition where if payloads are contain structs such as protobuf structs, the lack of a in-memory guard throws a lot of race conditions between slice.Append and slice.Slice calls.

* history/memstore: guard memstore read or modify calls with a mutex to prevent race conditions

* history/store: correct auto-correct's comment corrections

* auto/history: refactor to allow pollOnSchedule lifecycle defined by a cron schedule in config (#407)

* util/jsontypes: add a more relaxed cron parser

* auto/history: refactor to allow pollOnSchedule lifecycle defined by a cron schedule in config

* auto/history: default comparator in deduper constructor

* auto/history: use more concise for range x; remove unnecessary var declarations