ui/space: swap to infinite-scroll to render setup page; fix layout when airQuality gauge is missing; prevent nullPointer exceptions in airTemperature (#439)
The use of v-intersect was broken - if only a single zone was returned, the v-intersect would never leave the viewport causing the load more zone functionality to be permanently disabled until someone resized the viewport so the v-intersect leaves the viewport. The v-infinite-scroll component is a more reliable way to render zones and fetch more on scroll-to-bottom.

A fix for the airQuality component which would render the metric success/error status outside of the component if the airQuality gauge data was missing from the backend due to device limitations.

Prevent null pointer exceptions when the devices which support airTemperature trait don't return a value celsius reading.

* ui/space: swap to infinite-scroll to render setup page; fix layout when airQuality gauge is missing; prevent nullPointer exceptions in airTemperature

* ui/space: keep existing logic to check account zones first