system/gateway: remove ServicesApi special handling
The gateway used to have special handling to set up proxying for the ServicesApi, this was here to allow the legacy "automations", etc. names to be proxied by the gateway as "{node}/{service type}", something needed as the gateway had its own root names already. A while ago we changed how services were announced such that each node would announce the "{node}/{type}" name themselves, we also made it so that these names were advertised via the nodes ParentApi making them discoverable.

At the time we never updated the gateway code to reflect this change, leaving it so that we could support cohorts where some nodes were still using the old "{type}" pattern exclusively. I think we're past that point now. Keeping the old code would just result in errors in the logs as service types would be announced twice, once by the special handling and once when discovered via the nodes children.

This change removes (most of) the special handling of the ServicesApi. The names associated with the ServicesApi for a node are now announced along with all the other names a node advertises.

Special handling for gateways proxying gateways is needed. Normally a gateway would ignore all names advertised by another gateway, assuming that these will be discovered when querying the originating node. ServicesApi names are special though as a gateway can be the author of these names, requiring other gateways to proxy them. I expect this last special case will be removed when "name ownership" becomes a concept exposed by the api.