system/gateway/internal/rx: make writes block waiting for listeners
The previous behaviour was causing inconsistencies in the Sub methods, the value and chan did not agree on the state they returned causing duplicates in the best case, and potential data loss in the worst case.

Making the write synchronous makes deadlocks (caused by bad code) more likely, but we should be able to keep on top it.