system/gateway: use DevicesApi instead of ParentApi+MetadataApi for each child
This drastically reduces the network and goroutine usage. Before we'd need one routine and one pull stream for each node, then another goroutine and pull for each child on each node. Now we only need one per node. This will reduce the scale from ~10,000 down to ~10, a marked improvement, though I haven't benchmarked or tested this.

As part of this change, announcing the remote node itself no longer needs special-casing as the DevicesApi responds with the node and all its children. We still need to track the remote node for gateway processing, but hopefully that will go away once we know the host node of a device.

I've refactored the model to swap fields/types of child[ren] for device[s] as this now more accurately reflects what the data is modelling.