node: consistent locking for parent and metadata writes
While the underlying types used for routing, parent, and metadata modelling are thread safe, they don't behave consistently with each other, nor do they react to concurrent writes the way we want. A while ago we fixed this by locking around the Announce function, but we never completed that work.

This change adds locks to all writes to parent and metadata, updates the comment to make it clear what the lock is for, and deletes the old concurrency error checking code now we don't need it.

The lock isn't needed for reads as the router, parent, and metadata models are already safe for this kind of use.