node: remove NoAddChildTrait
I found no usages (outside of tests) in all projects using GitHub search.

This is preparation for making the ParentApi and the DevicesApi much more consistent. With this option it's possible to construct a node that returns "foo" from DevicesApi but not from the ParentApi:

```go
n.Announce("foo", NoAddChildTrait(), WithMetadata(md))
```

Note, the DevicesApi and ParentApi will always disagree on one name as the DevicesApi contains the node, the ParentApi does not. But we can still make it better.

A future change will reimplement the node's ParentApi on top of the existing metadata collection, instead of maintaining a separate collection of children.