internal/manage/devices: prep work to support non-string queries
I've reimplemented the scanning and query logic to allow non-string results to be collected and compared. While this change doesn't directly support non-string queries in the API, the foundation is there. A future change will adjust the DevicesApi to support more condition types.

As part of this I've significantly increased the test coverage for the query function, we now have over 90% coverage for the query functionality.

The basic approach is still the same, though I've moved a few bits around to make the logic of scanning a bit easier to follow as we descend into messages, lists, and maps. All query codepaths go through two branches: path-less, which uses protorange to visit every populated field, and path scanning, which visit fields matched by a path. The latter funcs are prefixed with `scan`.

The new code is more restrictive on what can be queried, now only leaf fields can be queried. This was implied before as only leaf fields could be converted to string. The new code makes `leaf` a documented concept which includes all scalar types, along with Timestamp and Duration. Additional leaf types may be added in the future.