node: store device information in a devicespb.Collection (#480)
The devices announced on a Node are presented to clients, both network and in-process, in three ways. Via children of the ParentApi when using the nodes name, via the MetadataApi, and via the DevicesApi. The MetadataApi and DevicesApi were both backed by a metadatapb.Collection, the ParentApi was backed by a parentpb.Model.

This change moves the backing store for all these into a (new) devicespb.Collection adjusting or introducing api servers that are now backed directly by this type.

For the DevicesApi the server got a bit simpler as conversion from a metadata collection to devices collection no longer needs to happen. As a side effect I've also adjusted the headers for the devices download feature to change "md.name" to just "name" now that it's backed with a Device instead of purely based on Metadata.

For the MetadataApi and ParentApi I've introduced some internal service impls that back off to various apis the Node now has for accessing devices.

Finally, I've updated usage of Node.ListAllMetadata to use the new ListDevices api instead of deprecating it. I couldn't find any use in dependent projects.