auth/token: fix enum attributes being ints
During serialisation the enum is encoded as an int, when it should be a string. We can't use protojson here, because that only has proto.Message API and nothing for enum types.

Also update the README to include the --v0-compatibility flag, as we need that now to run the opa tests.