sysconf: support setting port in grpcAddr / httpAddr
Sometimes it is necessary to override the port used for generating URLs, e.g. when running under Docker with port remapping.

In this case the listening port is different to the port that clients will access the service on.

Support host:port or :port in these fields, while maintaining the same behavior when just host is specified, and the auto-discovered host when nothing is specified.