ui/ops: allow customising or removing the title in EnvironmentalCard
In some dashboards the 'Environmental' title is not the best choice for this card.

Add a prop title, defaulting to 'Environmental', to customise the title.
If the title is set to an empty string, the title is not displayed at all (so there's no blank space at the top of the card).