proto,healthpb: add HealthApi and associated healthpb package
HealthApi is a trait API that associates one or more health checks against a device. This will eventually replace the Status trait and other supporting APIs like AlertApi and LightTest. See health.proto for details of use.

The healthpb package contains types and funcs for working with HealthChecks, the core type is registry.go#Registry which allows for the creation and management of health checks. Each type of health check (Fault, Bounds, etc) has its own healthpb type encapsulating the rules associated with checks of that type. For example the *BoundsCheck type updates a checks normality based on a value reading and configured normal bounds.

The healthmerge package is internal as a future PR will make use of it from another internal package.

The standard package contains one standard for use with health check as an example, more standards can be added later. Checks can be created with standards not in this list, though having them in one place can help with consistency.