tools/device-load-test: add tool to simulate the server load of a large floorplan
To help in diagnosing performance problems when the Ops UI has
a layered graphic referencing many devices, create a tool that
approximates the request pattern of the Ops UI when loading
such a graphic.

This allows us to evaluate the server performance separately
from the frontend performance, helping to identify the location
of any bottlenecks.