devices: add Query.Condition.matches
The new matches option for device queries allows a caller to say: return devices where this repeated item matches all these conditions. This was impossible before as conditions were matched against the Device only.

As an example, say you want to find devices that have metadata NICs with IP=1.2.3.4 and Network=1.2.3.0/24. This was impossible before, you were limited to a query that would return devices that had a NIC with the given IP and a NIC with the given Network, which might be different NICs.