devices: metadata now counts devices once per unique field value
Before, a device with a repeated field that contained more than one entry with the same value for a matched field would be counted twice in the field count. Now it will only be counted once.

Note that if a device has repeated fields that contain different values for a field, it will contribute to all unique values in the field counts.