ci: run Go tests on PRs that don't target main
At present, Go test actions are only run for PRs that have `main` as their base branch, which isn't the case when using stacked PRs.

Change the triggers for the Go Test action to run for all PRs. To reduce the number of unnecessary action runs, only run Go tests for changes that have touched the Go parts of the project.