system/gateway/internal/test: decrease max backoff and log timing
This test is being flaky, add log messages so we can track (some of) where time is being spent on the assumption that things are working fine but the test isn't waiting long enough. Reduce the max backoff interval from 60s to 5s to make sure we're not going to be competing with the 60s deadline too much.