app,proto: remove measured values from DevicesApi health checks
With the addition of health checks to Device the frequency of updates via APIs like PullDevices has increased by orders of magnitude without really adding any value to use cases that use the DevicesApi. This removes "measured values" from the devices api health checks as these are the primary source of non-useful updates via this API.

The code feels a little disjointed from the DevicesApi server because it's tied to the write operation, rather than the read.

I've also enabled deduplication for pulls via the DevicesApi as the implementation approach here relies on it to avoid sending devices with identical health checks twice in a row.