devices: add DevicesApi.Query.Condition.matcher
This new field allows you to specify how a condition behaves when matching against repeated fields. For example, a condition like `{stringContains: "bana"}` using the ANY (or default) matcher setting would match against a device with these resolved values: ["banana", "apple"], using ALL would not match this device as "apple" doesn't contain "bana".

There is room in the Matcher enum for a NOT matcher, which is something we've been after for a while. This change doesn't include it as it's not strictly needed for this.